Client network

+++ Engine/includes/Nexus/frameworks/NetworkServerFrameworkTpl.hpp
@@ -21,7 +21,7 @@ namespace nx {
 		// Framework Methods (must be virtual pure)
 		virtual void sendEvent(const nx::netserv_send_event_t& netInfos) = 0;
 		virtual void sendAll(const nx::netserv_send_event_t& net) = 0;
-		virtual void disconnect(const uint8_t clientId) = 0;
+		virtual void disconnect(const int clientId) = 0;
 		virtual void connectClient(const nx::netserv_client_t& clientInfos) = 0;
 	};
 }
